pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "overwatch"
version = "0.1.0"
description = "TUI process monitor with pluggable monitors and SDK"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.50.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "httpx>=0.25",
]

[project.scripts]
overwatch = "overwatch.cli:main"

[tool.setuptools.packages.find]
include = ["overwatch*", "sdk*"]
